Move remaining version endpoints to version::*
